update new pkgInfo action

+++ cangjie-language-server/src/languageserver/ArkServer.cpp
@@ -925,7 +925,9 @@ void ArkServer::ChangeWatchedFiles(const std::string &file, FileChangeType type,
         if (type == FileChangeType::DELETED) {
             Logger::Instance().LogMessage(MessageType::MSG_INFO, "delete the file:  " + file);
             CompilerCangjieProject::GetInstance()->IncrementForFileDelete(file);
-            CompilerCangjieProject::GetInstance()->GetBgIndexDB()->DeleteFiles({file});
+            if (CompilerCangjieProject::GetUseDB()) {
+                CompilerCangjieProject::GetInstance()->GetBgIndexDB()->DeleteFiles({file});
+            }
             this->callback->RemoveDocByFile(input.inputs.fileName);
         }
         if (!FileUtil::FileExist(input.onEditFile)) {
